# spec/edited/openapi.yaml
openapi: 3.0.2
info:
  title: Wellbore DDMS OSDU
  description: build local
  version: '0.2'
paths:
...
  '/ddms/v2/logsets/{logsetid}':
    get:
      tags:
        - Logset
      summary: 'Get the LogSet using wks:logSet:1.0.5 schema'
      description: Get the LogSet object using its **id**
      operationId: get_logset
      parameters:
        - required: true
          schema:
            title: Logsetid
            type: string
          name: logsetid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/logset'
        '404':
          description: LogSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    delete:
      tags:
        - Logset
      summary: Delete the LogSet. The API performs a logical deletion of the given record
      operationId: del_logset
      parameters:
        - required: true
          schema:
            title: Logsetid
            type: string
          name: logsetid
          in: path
        - description: Whether or not to delete records children
          required: false
          schema:
            title: Recursive
            type: boolean
            description: Whether or not to delete records children
            default: false
          name: recursive
          in: query
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '204':
          description: Record deleted successfully
        '404':
          description: LogSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/logsets/{logsetid}/versions':
    get:
      tags:
        - Logset
      summary: Get all versions of the logset
      operationId: get_logset_versions
      parameters:
        - required: true
          schema:
            title: Logsetid
            type: string
          name: logsetid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordVersions'
        '404':
          description: LogSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/logsets/{logsetid}/versions/{version}':
    get:
      tags:
        - Logset
      summary: 'Get the given version of LogSet using wks:logSet:1.0.5 schema'
      description: '"Get the LogSet object using its **id**.'
      operationId: get_logset_version
      parameters:
        - required: true
          schema:
            title: Logsetid
            type: string
          name: logsetid
          in: path
        - required: true
          schema:
            title: Version
            type: integer
          name: version
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/logset'
        '404':
          description: LogSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
...
  '/ddms/v2/markers/{markerid}':
    get:
      tags:
        - Marker
      summary: 'Get the marker using wks:marker:1.0.4 schema'
      description: Get the Marker object using its **id**
      operationId: get_marker
      parameters:
        - required: true
          schema:
            title: Markerid
            type: string
          name: markerid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/marker'
        '404':
          description: marker not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    delete:
      tags:
        - Marker
      summary: Delete the marker. The API performs a logical deletion of the given record
      operationId: del_marker
      parameters:
        - required: true
          schema:
            title: Markerid
            type: string
          name: markerid
          in: path
        - description: Whether or not to delete records children
          required: false
          schema:
            title: Recursive
            type: boolean
            description: Whether or not to delete records children
            default: false
          name: recursive
          in: query
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '204':
          description: Record deleted successfully
        '404':
          description: Marker not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/markers/{markerid}/versions':
    get:
      tags:
        - Marker
      summary: Get all versions of the marker
      operationId: get_marker_versions
      parameters:
        - required: true
          schema:
            title: Markerid
            type: string
          name: markerid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordVersions'
        '404':
          description: marker not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/markers/{markerid}/versions/{version}':
    get:
      tags:
        - Marker
      summary: 'Get the given version of marker using wks:marker:1.0.4 schema'
      operationId: get_marker_version
      parameters:
        - required: true
          schema:
            title: Markerid
            type: string
          name: markerid
          in: path
        - required: true
          schema:
            title: Version
            type: integer
          name: version
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/marker'
        '404':
          description: marker not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  /ddms/v2/markers:
    post:
      tags:
        - Marker
      summary: 'Create or update the markers using wks:marker:1.0.4 schema'
      operationId: post_marker
      parameters:
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              title: Markers
              type: array
              items:
                $ref: '#/components/schemas/marker'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUpdateRecordsResponse'
        '400':
          description: Missing mandatory parameter or unknown parameter
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
...
components:
  schemas:
...
    CreateUpdateRecordsResponse:
      title: CreateUpdateRecordsResponse
      type: object
      properties:
        recordCount:
          title: Recordcount
          type: integer
        recordIds:
          title: Recordids
          type: array
          items:
            type: string
        skippedRecordIds:
          title: Skippedrecordids
          type: array
          items:
            type: string
    DataType:
      title: DataType
      enum:
        - string
        - number
        - integer
        - boolean
      description: An enumeration.
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        errors:
          title: Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    Kind:
      title: Kind
      enum:
        - CRS
        - Unit
        - Measurement
        - AzimuthReference
        - DateTime
      type: string
      description: An enumeration.
    LinkList:
      title: LinkList
      type: object
      properties: {}
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    MetaItem:
      title: MetaItem
      required:
        - kind
        - persistableReference
      type: object
      properties:
        kind:
          title: Reference Kind
          allOf:
            - $ref: '#/components/schemas/Kind'
          description: 'The kind of reference, unit, measurement, CRS or azimuth reference.'
        name:
          title: Name or Symbol
          type: string
          description: The name of the CRS or the symbol/name of the unit
        persistableReference:
          title: Persistable Reference
          type: string
          description: The persistable reference string uniquely identifying the CRS or Unit
        propertyNames:
          title: Attribute Names
          type: array
          items:
            type: string
          description: 'The list of property names, to which this meta data item provides Unit/CRS context to. Data structures, which come in a single frame of reference, can register the property name, others require a full path like "data.structureA.propertyB" to define a unique context.'
        propertyValues:
          title: Attribute Names
          type: array
          items:
            type: string
          description: 'The list of property values, to which this meta data item provides Unit/CRS context to. Typically a unit symbol is a value to a data structure; this symbol is then registered in this propertyValues array and the persistableReference provides the absolute reference.'
        uncertainty:
          title: Uncertainty
          type: number
          description: The uncertainty of the values measured given the unit or CRS unit.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    RecordVersions:
      title: RecordVersions
      type: object
      properties:
        recordId:
          title: Recordid
          type: string
        versions:
          title: Versions
          type: array
          items:
            type: integer
...
    SimpleElevationReference:
      title: SimpleElevationReference
      required:
        - elevationFromMsl
      type: object
      properties:
        elevationFromMsl:
          title: Elevation from MSL
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The elevation above mean sea level (MSL), at which the vertical origin is 0.0. The ''unitKey'' is further defined in ''frameOfReference.units''.'
        name:
          title: Elevation Reference Name
          type: string
          description: The name of the Elevation Reference.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    TagDictionary:
      title: TagDictionary
      type: object
      properties: {}
      description: Used for data model allows extra fields which are not declared initially in the pydantic model
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    ValueWithUnit:
      title: ValueWithUnit
      required:
        - unitKey
        - value
      type: object
      properties:
        unitKey:
          title: Unit Key
          type: string
          description: Unit for value of the corresponding attribute for the domain object in question. The key can be looked up in the 'frameOfReference.units' for further details.
        value:
          title: Value
          type: number
          description: Value of the corresponding attribute for the domain object in question.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    app__model__model_curated__Legal:
      title: Legal
      type: object
      properties:
        legaltags:
          title: Legal Tags
          type: array
          items:
            type: string
          description: 'The list of legal tags, see compliance API.'
        otherRelevantDataCountries:
          title: Other Relevant Data Countries
          type: array
          items:
            type: string
          description: 'The list of other relevant data countries using the ISO 2-letter codes, see compliance API.'
        status:
          title: Legal Status
          type: string
          description: The legal status.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    channel:
      title: channel
      type: object
      properties:
        absentValue:
          title: Absent Value
          type: string
          description: Optional field carrying the absent value as string for this channel.
        dataType:
          title: Data Type
          allOf:
            - $ref: '#/components/schemas/DataType'
          description: The log value type (per log sample). The 'format' property may contain further hints about data type presentation.
          default: number
        dimension:
          title: Dimension
          type: integer
          description: The dimension of this log or channel
        family:
          title: Log Family
          type: string
          description: The log family code of this log or channel (optional)
        familyType:
          title: Log Family Type
          type: string
          description: 'The log family type code of this log or channel. Example: ''Neutron Porosity'' for ''Thermal Neutron Porosity Sandstone''. (optional)'
        format:
          title: Format Hint
          allOf:
            - $ref: '#/components/schemas/Format'
          description: Optional format hint how to treat the log values as strings or number of bits per 'dataType'.
          default: float32
        logstoreId:
          title: Logstore ID
          type: number
          description: The id of this log or channel in the Logstore. This property is not present in the index channel.
        bulkURI:
          title: bulk URI
          type: string
          description: bulkURI either URL or URN.
        longName:
          title: Log Long Name
          type: string
          description: The long name of this log or channel
        mnemonic:
          title: Mnemonic
          type: string
          description: The mnemonic of this log or channel
        name:
          title: Log Name
          type: string
          description: The name of this log or channel.
        properties:
          title: Named Properties
          type: array
          items:
            $ref: '#/components/schemas/namedProperty'
          description: The named properties of this log or channel.
        source:
          title: Source
          type: string
          description: 'The source of this log or channel as a data reference; Typically this refers to the raw LogSet, from which this log WKE is generated.'
        unitKey:
          title: Unit
          type: string
          description: The unitKey to be looked up in the 'frameOfReference.units' dictionary to find the self-contained definition.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    logSetData:
      title: logSetData
      type: object
      properties:
        azimuthReference:
          title: Azimuth Reference Code
          type: string
          description: Azimuth reference code defining the type of North. Only used for logSets with azimuth data
        channelMnemonics:
          title: Channel Mnemonics
          type: array
          items:
            type: string
          description: A list of channel Mnemonics in this log set.
        channelNames:
          title: Channel Names
          type: array
          items:
            type: string
          description: A list of channel long names in this log set.
        classification:
          title: Log Set Classification
          type: string
          description: The well-known log set classification code.
          default: Externally Processed LogSet
        dateCreated:
          title: Creation Date and Time
          type: string
          description: The UTC date time of the entity creation
          format: date-time
        dateModified:
          title: Last Modification Date and Time
          type: string
          description: The UTC date time of the last entity modification
          format: date-time
        elevationReference:
          $ref: '#/components/schemas/SimpleElevationReference'
        externalIds:
          title: Array of External IDs
          type: array
          items:
            type: string
          description: 'An array of identities (e.g. some kind if URL to be resolved in an external data store), which links to external realizations of the same entity.'
        name:
          title: Log Set Name
          type: string
          description: The name of this log set
        operation:
          title: Operation
          type: string
          description: The operation which created this entity
        reference:
          $ref: '#/components/schemas/channel'
        referenceType:
          title: Reference Type
          type: string
          description: The reference index type of the log set.
        relationships:
          $ref: '#/components/schemas/logsetrelationships'
        start:
          $ref: '#/components/schemas/ValueWithUnit'
        step:
          $ref: '#/components/schemas/ValueWithUnit'
        stop:
          $ref: '#/components/schemas/ValueWithUnit'
      description: Used for data model allows extra fields which are not declared initially in the pydantic model
...
    logset:
      title: logset
      type: object
      properties:
        acl:
          title: Access Control List
          allOf:
            - $ref: '#/components/schemas/TagDictionary'
          description: The access control tags associated with this entity.
        ancestry:
          title: Ancestry
          allOf:
            - $ref: '#/components/schemas/LinkList'
          description: 'The links to data, which constitute the inputs.'
        data:
          title: Log Set Data
          allOf:
            - $ref: '#/components/schemas/logSetData'
          description: Log channel set associated with a wellbore
        id:
          title: Log Set ID
          type: string
          description: The unique identifier of the log set
        kind:
          title: Log Set Kind
          type: string
          description: Kind specification
          default: 'osdu:wks:logSet:0.0.1'
        legal:
          title: Legal Tags
          allOf:
            - $ref: '#/components/schemas/app__model__model_curated__Legal'
          description: The log-set's legal tags
        meta:
          title: Frame of Reference Meta Data
          type: array
          items:
            $ref: '#/components/schemas/MetaItem'
          description: 'The meta data section linking the ''unitKey'', ''crsKey'' to self-contained definitions (persistableReference)'
        type:
          title: Entity Type
          type: string
          description: 'The reference entity type as declared in common:metadata:entity:*.'
        version:
          title: Entity Version Number
          type: number
          description: The version number of this log set; set by the framework.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
    logsetrelationships:
      title: logsetrelationships
      required:
        - wellbore
      type: object
      properties:
        well:
          title: Well
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The well to which this logSet belongs. Only required if the wellbore is unknown.
        wellbore:
          title: Wellbore
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The wellbore to which this logSet belongs.
        wellboreSection:
          title: Wellbore Section
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The wellboreSection to which this logSet belongs.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
    marker:
      title: marker
      required:
        - acl
        - kind
        - legal
      type: object
      properties:
        acl:
          title: Access Control List
          allOf:
            - $ref: '#/components/schemas/TagDictionary'
          description: The access control tags associated with this entity.
        ancestry:
          title: Ancestry
          allOf:
            - $ref: '#/components/schemas/LinkList'
          description: 'The links to data, which constitute the inputs.'
        data:
          title: Marker Data
          allOf:
            - $ref: '#/components/schemas/markerData'
          description: 'Geological marker using a single point-observation, typically along a wellbore.'
        id:
          title: Marker ID
          type: string
          description: The unique identifier of the marker
        kind:
          title: Marker Kind
          type: string
          description: Marker kind specification
        legal:
          title: Legal Tags
          allOf:
            - $ref: '#/components/schemas/app__model__model_curated__Legal'
          description: The marker's legal tags
        meta:
          title: Frame of Reference Meta Data
          type: array
          items:
            $ref: '#/components/schemas/MetaItem'
          description: 'The meta data section linking the ''unitKey'', ''crsKey'' to self-contained definitions (persistableReference)'
        type:
          title: Entity Type
          type: string
          description: 'The reference entity type as declared in common:metadata:entity:*.'
        version:
          title: Entity Version Number
          type: number
          description: The version number of this marker; set by the framework.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
  securitySchemes:
    OpenDESBearerToken:
      type: http
      scheme: bearer